+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - ISelectable interface
 - ISubmitable interface
 
+### Changed
+- DelayedButton implements ISelectable and ISubmitable interfaces
+
 ## [3.0.0] - 2021-12-26
 ### Changed
 - BaseMenu: Replace Buttons by MenuEventTriggers